split boundless to its own stat. remove overhealing
